FaceCount-ClientApplication/requirements.txt
pillow
requests
numpy
opencv-python
face_recognition
cryptography 